+++ changelog.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Do not use Methods as Singleton
+- Accept null as value for some fields via REST API
 
 ## [3.0.1] - 02.08.2026